+++ tests/test_modules/test_constraints.py
@@ -243,7 +243,7 @@ class TestConstraintsMixin:
         assert "exceeds number of" in str(excinfo.value)
 
         with pytest.raises(AssertionError) as excinfo:
-            scwc.constrain(np.int_(100))
+            scwc.constrain(np.int32(100))
         assert "exceeds number of" in str(excinfo.value)
 
     def test_constrain_with_unregistered_system_throws(
